switch to es6 modules to make code clearer
Whew!  Now I can see the structure of the code quite
a bit better.  Looks like js/ecl.js was entirely unused,
js/ecl/main.js is used for *a lot of things,* and there's
a couple of things used by scripts embedded in MD/HTML code
that I've placed as variables on `window` for now.