add scripts for generating anm-stats.json
These were formerly just sitting around in my thcrap folder
not in version control.... and when I first found the anm-usage.py
script today, it was entirely broken, halfway in the middle of a
huge refactor with a bunch of half-implemented features!
Again, not in version control!  What was I thinking?!?!