compress anm-stats.json
This prepares the website for adding BM stats by:

- Sorting the data in anm-stats.json by opcode.  (previously, the order
  was *deterministic* but kind of arbitrary)
- Doing sort of a dumb looking thing where it performs bz2 compression
  on some of the JSON objects inside anm-stats.json, writing them as
  base64-encoded blobs instead.

Originally I was going to replace "num-files" with a "files": [list
of filenames] and then put file indices in the breakdowns, but this
resulted in the data looking pretty cryptic like [[4, 20], [6, 9]]
and it only produced about a 50% reduction in size IIRC.

After this I thought, *well gee*, if these handwritten compression
schemes have so little benefit and still render the data unreadable, I
might as well just use general compression.  But it's done to only
portions of the file, so that we can keep adding more stuff to the file
without disturbing the existing data in our git diffs.

....hence, base64-encoded, bz2-compressed json-ception.

(this ended up being more difficult than anticipated to decompress
back on the JS end, but eventually that got ironed out).